replace with simple navbar, move updated text to about page
