+++ ampere/app.py
@@ -1,10 +1,7 @@
-import datetime
-
 import dash
 import dash_bootstrap_components as dbc
-from dash import html, Input, Output, State
+from dash import html
 
-from ampere.common import get_db_con
 from ampere.styling import AmperePalette
 
 app = dash.Dash(
@@ -12,106 +9,46 @@ app = dash.Dash(
     external_stylesheets=[dbc.themes.BOOTSTRAP],
 )
 
-
-def get_last_updated() -> datetime.datetime:
-    con = get_db_con()
-    max_retrieved_at = (
-        con.sql("SELECT max(retrieved_at) as last_updated FROM main.repos")
-        .to_df()
-        .squeeze()
-    )
-
-    return max_retrieved_at
-
-
-last_updated = get_last_updated()
-last_updated_str = last_updated.strftime("%Y-%m-%d")
-page_links_collapsed = dbc.Row(
-    [
-        dbc.Col(
-            dbc.DropdownMenu(
-                children=[
-                    dbc.DropdownMenuItem(
-                        "stargazers",
-                        href="network-stargazers",
-                        style={"color": "black"},
-                    ),
-                    dbc.DropdownMenuItem(
-                        "followers", href="network-followers", style={"color": "black"}
-                    ),
-                ],
-                label="networks",
-                toggle_style={
-                    "color": AmperePalette.BRAND_TEXT_COLOR_MUTED,
-                    "backgroundColor": AmperePalette.PAGE_ACCENT_COLOR,
-                    "borderColor": AmperePalette.PAGE_ACCENT_COLOR,
-                },
-                class_name="navbar-text",
-            ),
+navbar = dbc.NavbarSimple(
+    children=[
+        dbc.DropdownMenu(
+            children=[
+                dbc.DropdownMenuItem(
+                    "stargazers",
+                    href="network-stargazers",
+                    style={"color": "black"},
+                ),
+                dbc.DropdownMenuItem(
+                    "followers", href="network-followers", style={"color": "black"}
+                ),
+            ],
+            nav=True,
+            in_navbar=True,
+            label="networks",
         ),
-        dbc.Col(
-            dbc.NavItem(
-                dbc.NavLink(
-                    "about",
-                    href="about",
-                    style={"color": AmperePalette.BRAND_TEXT_COLOR_MUTED},
-                    class_name="navbar-text",
-                )
-            ),
+        dbc.NavItem(
+            dbc.NavLink(
+                "about",
+                href="about",
+                style={"color": AmperePalette.BRAND_TEXT_COLOR_MUTED},
+                class_name="navbar-text",
+            )
         ),
     ],
-    align="left",
-    class_name="row align-items-center",
-)
-
-navbar = dbc.Navbar(
-    dbc.Container(
-        [
-            html.A(
-                dbc.Row([dbc.NavbarBrand("ampere")]),
-                href="/",
-                style={"textDecoration": "none"},
-            ),
-            dbc.NavbarToggler(id="navbar-toggler", n_clicks=0),
-            dbc.Collapse(
-                page_links_collapsed,
-                id="navbar-collapse",
-                is_open=False,
-                navbar=True,
-            ),
-            html.P(
-                f"last updated {last_updated_str}",
-                style={
-                    "marginBottom": "0",
-                    "color": AmperePalette.BRAND_TEXT_COLOR_MUTED,
-                    "fontSize": "12px",
-                },
-            ),
-        ],
-        fluid=True,
-    ),
     color=AmperePalette.PAGE_ACCENT_COLOR,
     dark=True,
     fixed="top",
-    style={"padding": "0"},
+    fluid=True,
+    style={"height": "5%", "width": "100%"},
+    brand = "ampere",
+    links_left=True,
+    brand_href="/",
 )
-
-
-# add callback for toggling the collapse on small screens
-@app.callback(
-    Output("navbar-collapse", "is_open"),
-    [Input("navbar-toggler", "n_clicks")],
-    [State("navbar-collapse", "is_open")],
-)
-def toggle_navbar_collapse(n, is_open):
-    if n:
-        return not is_open
-    return is_open
-
 
 app.layout = dbc.Container(
     [
         navbar,
+        html.Br(),
         html.Br(),
         dash.page_container,
     ],
